Move UiaClient interop into the test project
